backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic>=2.9.0
web3==6.11.3
solana==0.30.2
slither-analyzer==0.10.0
reportlab==4.0.7
python-multipart==0.0.6
aiofiles==23.2.1
openai==1.3.7
python-dotenv==1.0.0
httpx>=0.23.0,<0.24.0
py-solc-x==1.1.1
numpy>=1.24.0
aiohttp>=3.9.0
setuptools>=69.0.0
requests>=2.31.0

